+++ docs/changes.md
@@ -1,5 +1,6 @@
 2022/01/06
   * fscript中增加is\_int/is\_odd/is\_even等函数
+  * fscript增加random0to1。
 
 2022/01/05
   * fscript公开函数value\_id/value\_func。